+++ swen30006/automail/StorageTube.java
@@ -12,15 +12,42 @@ public class StorageTube {
 
     public final int MAXIMUM_CAPACITY;
     public Stack<MailItem> tube;
+    private IMailPool mailPool;
 
     /**
      * Constructor for the storage tube
      */
-    public StorageTube(int maxCap){
+    public StorageTube(int maxCap, IMailPool mailPool){
     	this.MAXIMUM_CAPACITY = maxCap;
         this.tube = new Stack<MailItem>();
+        this.mailPool = mailPool;
     }
 
+    /**
+     * @param strong is whether the tube belongs to a strong robot.
+     */
+	void fillStorageTube(boolean strong) {
+		/* Determine max weight for the tube via the properties file */
+		int maxWeight = strong ? Integer.parseInt(MyProps.getProp("Max_Weight")) : Integer.parseInt(MyProps.getProp("Weak_Max_Weight"));
+		MailItem curItem;
+		
+		try{
+			while(!isFull()) {
+				curItem = mailPool.getMail(maxWeight);
+				if(curItem != null) {
+					this.addItem(curItem);  // Could group/order by floor taking priority into account - but already better than simple
+					
+				} else {
+					break;
+				}
+			}
+		}
+		catch(TubeFullException e){
+			e.printStackTrace();
+			
+		}
+	}
+    
     /**
      * @return if the storage tube is full
      */